requirements.txt
dj-database-url==0.5.0
Django==2.1.1
django-extensions==2.1.3
django-filter==2.0.0
django-heroku==0.3.1
djangorestframework==3.8.2
gunicorn==19.9.0
newsapi-python
Cython
git+git://github.com/networkx/networkx-metis.git